requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
npm==0.1.1
optional-django==0.1.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
